refactor: study timer 폴더로 묶음 + 주석 지우기
