+++ app/src/main/res/layout/fragment_study_countdown.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".StudyFragment2">
+    tools:context=".STUDY.StudyCountDownFragment">
 
 
     <LinearLayout
