+++ app/src/main/res/layout/fragment_study_countup.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StudyFragment">
+    tools:context=".STUDY.StudyCountUpFragment">
 
     <!-- TODO: Update blank fragment layout -->
